+++ anta/cli/check/__init__.py
@@ -1,7 +1,9 @@
 # Copyright (c) 2023 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
-
+"""
+Click commands to validate configuration files
+"""
 import click
 
 from anta.cli.check import commands
