+++ anta/cli/debug/__init__.py
@@ -1,6 +1,9 @@
 # Copyright (c) 2023 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
+"""
+Click commands to execute EOS commands on remote devices
+"""
 import click
 
 from anta.cli.debug import commands
