+++ anta/cli/exec/__init__.py
@@ -1,13 +1,16 @@
 # Copyright (c) 2023 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
+"""
+Click commands to execute various scripts on EOS devices
+"""
 import click
 
 from anta.cli.exec import commands
 
 
 @click.group
-def exec() -> None:
+def exec() -> None:  # pylint: disable=redefined-builtin
     """Execute commands to inventory devices"""
