+++ anta/cli/exec/commands.py
@@ -2,7 +2,7 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 """
-Commands for Anta CLI to execute EOS commands.
+Click commands to execute various scripts on EOS devices
 """
 from __future__ import annotations
