+++ anta/cli/nrfu/commands.py
@@ -2,7 +2,7 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 """
-Commands for Anta CLI to run nrfu commands.
+Click commands that render ANTA tests results
 """
 from __future__ import annotations
